+++ ListlyApp/src/constants/icons.ts
@@ -1,0 +1,15 @@
+import type { IconName } from './types';
+
+export const ICONS: {
+  collection: IconName;
+  list: IconName;
+  edit: IconName;
+  notFound: IconName;
+  removeFromCollection: IconName;
+} = {
+  collection: 'albums-outline',
+  list: 'list-outline',
+  edit: 'create-outline',
+  notFound: 'help-circle-outline',
+  removeFromCollection: 'arrow-undo-outline',
+};